test: add 20.0-project E2E test asset and update update tests to use it
Update e2e assets
