+++ tests/e2e/assets/20.0-project/.editorconfig
