+++ tests/e2e/assets/20.0-project/README.md
@@ -1,6 +1,6 @@
-# NineteenProject
+# TwentyProject
 
-This project was generated using [Angular CLI](https://github.com/angular/angular-cli) version 19.2.13.
+This project was generated using [Angular CLI](https://github.com/angular/angular-cli) version 20.3.27.
 
 ## Development server
