+++ tests/e2e/tests/update/update.ts
@@ -12,10 +12,10 @@ export default async function () {
   try {
     // We need to use the public registry because in the local NPM server we don't have
     // older versions @angular/cli packages which would cause `npm install` during `ng update` to fail.
-    restoreRegistry = await createProjectFromAsset('19.0-project', true);
+    restoreRegistry = await createProjectFromAsset('20.0-project', true);
 
     // CLI project version
-    const cliMajorProjectVersion = 19;
+    const cliMajorProjectVersion = 20;
 
     // If using npm, enable legacy peer deps mode to avoid defects in npm 7+'s peer dependency resolution
     // Example error where 11.2.14 satisfies the SemVer range ^11.0.0 but still fails:
@@ -72,8 +72,8 @@ export default async function () {
   await ng('update', '@angular/cli', ...extraUpdateArgs);
 
   // Setup testing to use CI Chrome.
-  await useCIChrome('nineteen-project', './');
-  await useCIDefaults('nineteen-project');
+  await useCIChrome('twenty-project', './');
+  await useCIDefaults('twenty-project');
 
   // Run CLI commands.
   await ng('generate', 'component', 'my-comp');
@@ -81,14 +81,14 @@ export default async function () {
 
   await executeBrowserTest({
     configuration: 'production',
-    expectedTitleText: 'Hello, nineteen-project',
+    expectedTitleText: 'Hello, twenty-project',
   });
   await executeBrowserTest({
     configuration: 'development',
-    expectedTitleText: 'Hello, nineteen-project',
+    expectedTitleText: 'Hello, twenty-project',
   });
 
   // Verify project now creates bundles
   await noSilentNg('build', '--configuration=production');
-  await expectFileMatchToExist('dist/nineteen-project/browser', /main-[a-zA-Z0-9]{8}\.js/);
+  await expectFileMatchToExist('dist/twenty-project/browser', /main-[a-zA-Z0-9]{8}\.js/);
 }
